requirements.txt
transformers
webrtcvad
build
torch
sounddevice
paho-mqtt